+++ src/ClientSession/ClientSession.hpp
@@ -1,9 +1,9 @@
 #ifndef CLIENT_SESSION_HPP
 #define CLIENT_SESSION_HPP
 
-#include "RequestParser.hpp"
-#include "GlobalConfig.hpp"
-#include "commonEnums.hpp"
+#include "../RequestParser/RequestParser.hpp"
+#include "../GlobalConfig/GlobalConfig.hpp"
+#include "../include/commonEnums.hpp"
 
 class ClientSession {
 	public:
