+++ src/RequestHandler/StaticHandler.hpp
@@ -1,11 +1,11 @@
 #ifndef STATIC_HANDLER_HPP
 #define STATIC_HANDLER_HPP
 
-#include "GlobalConfig.hpp"
-#include "ClientSession.hpp"
-#include "RequestMessage.hpp"
-#include "ResponseBuilder.hpp"
-#include "file_utils.hpp"
+#include "../GlobalConfig/GlobalConfig.hpp"
+#include "../ClientSession/ClientSession.hpp"
+#include "../RequestMessage/RequestMessage.hpp"
+#include "../ResponseBuilder/ResponseBuilder.hpp"
+#include "../utils/file_utils.hpp"
 #include <sstream>
 #include <vector>
 #include <string>
